api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
